Split resource repro modules for architecture gate
